feat: MaterialSystem overhaul
it wasn't done but it still got an overhaul. shaders, textures, and materials are now more tightly integrated with eachother, being combined into one module handling them all via materials. handles are still in renderer.

some veldrid code is inside of the matsys code but that is fine as the matsys is very rendering related.